# Improved performance of the safe prime generator